chore: use css variables for fonts
